% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_annotate.R
\name{add_zero_to_single}
\alias{add_zero_to_single}
\title{Adds a leading zero to a number if less than 10}
\usage{
add_zero_to_single(x)
}
\description{
Adds a leading zero to a number if less than 10
}
\examples{
add_zero_to_single(2)
}
