% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_annotate.R
\name{annotate_plate}
\alias{annotate_plate}
\title{Produces an annotation data frame given a plate map}
\usage{
annotate_plate(df, type = 96, plate = 1, area = 2, condition = 3,
  value = 4)
}
\arguments{
\item{df}{data frame containing \code{\link{wellrange}}s along with the condition and value to be applied.
For example, if row A contains a drug with two different concentrations and row B contains a control:
Plate    Area    Condition    Value
1    A1-A12    Drug    Treated    
1    B1-B12    Drug    Control
1    A1-A6    Concentration    10
1    A7-A12    Concentration    100}

\item{type}{numeric 96 or 384 depending on the type of plate to be annotated}

\item{plate}{numeric column index of Plate identifier column}

\item{area}{numeric column index of Area column (range of wells to be annotated)}

\item{condition}{numeric column index of Condition column}

\item{value}{numeric column index of Value column}
}
\value{
a data frame with 96 or 384 rows and columns for each Condition containing the value for that well.
This can be merged (joined) with a data frame of data values for wells for easy annotation.
}
\description{
Produces an annotation data frame given a plate map
}
