% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenamer.R
\name{filenamer}
\alias{filenamer}
\title{This function reads in a set of files, adds a column with the filename (with extension stripped) and combines them into one big tibble.}
\usage{
filenamer(pattern, path = "./", replacements = c(".xls", ".txt", ".tsv",
  ".csv", ".xlsx"), type = "tsv")
}
\arguments{
\item{pattern}{input character Pattern to match filenames (e.g. "*.xls").}

\item{path}{input character Path to directory to search for files.}

\item{replacements}{input character vector These patterns will be deleted from filenames to form the sample name.}

\item{type}{input character "tsv" or "excel" to specify whether \link{read_tsv} or \link{read_excel} is used to import data.}
}
\value{
A tibble combining all input files with a new column giving the sample (derived from stripping the 
replacements from the filename). If the files have different numbers of columns, a list of tibbles is returned instead.
}
\description{
This function reads in a set of files, adds a column with the filename (with extension stripped) and combines them into one big tibble.
}
\examples{
\dontrun{ filenamer("*.txt",path="~/data/", replacements=c(".txt"), type="tsv") }  
}
\seealso{
\code{\link{list.files}, \link{gsub}} which this function wraps.
}
