% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_annotate.r
\name{wellrange}
\alias{wellrange}
\title{Generates empty data frames given a well range}
\usage{
wellrange(range, first = "row")
}
\arguments{
\item{range}{character range of wells to include, start and end well separated by a dash e.g. A1-B12}

\item{first}{character "row" (default) runs row first, i.e A1, A2, A3 ... B1, B2 ... Any other value leads to
column-first output, i.e. A1, B1, C1 ... A2, B2, C2 ...}
}
\value{
Empty data frame with well numbers
}
\description{
Generates empty data frames given a well range
}
\examples{
wellrange("A1-B5")
wellrange("A1-B5",first="column")
}
